ew2025-image.sh: Build mesa only if not available
Building mesa with usermode emulation on x86_64 takes a long time.
Also the script assumes that the mesa sources are located in ~/src/mesa
and have the imx8mp-demo-ew2025 branch.

To avoid building mesa and make sure we get the sources from the correct
remote first check if a mesa tree is available already and assume it has
the correct refs. If the tree is not available clone it from Tomeu's
remote.

Finally rsync the mesa tree to the chroot.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>